+++ src/modules/tab_list/entry.rs
@@ -3,14 +3,14 @@ use std::os::raw::c_int;
 
 #[derive(Debug, Clone)]
 pub struct TabListEntry {
-  pub id: usize,
+  pub id: u8,
   pub real_name: String,
   pub nick_name: String,
   pub group: String,
 }
 
 impl TabListEntry {
-  pub fn from_id(id: usize) -> Self {
+  pub fn from_id(id: u8) -> Self {
     Self {
       id,
       real_name: tablist_get_real_name(id),
@@ -20,33 +20,33 @@ impl TabListEntry {
   }
 }
 
-fn tablist_get_real_name(id: usize) -> String {
+fn tablist_get_real_name(id: u8) -> String {
   // or "Player"
   unsafe {
     StringsBuffer_UNSAFE_Get(
       &mut TabList._buffer,
-      c_int::from(TabList.NameOffsets[id] - 3),
+      c_int::from(TabList.NameOffsets[id as usize] - 3),
     )
   }
   .to_string()
 }
 
-fn tablist_get_nick_name(id: usize) -> String {
+fn tablist_get_nick_name(id: u8) -> String {
   // or "Text"
   unsafe {
     StringsBuffer_UNSAFE_Get(
       &mut TabList._buffer,
-      c_int::from(TabList.NameOffsets[id] - 2),
+      c_int::from(TabList.NameOffsets[id as usize] - 2),
     )
   }
   .to_string()
 }
 
-fn tablist_get_group(id: usize) -> String {
+fn tablist_get_group(id: u8) -> String {
   unsafe {
     StringsBuffer_UNSAFE_Get(
       &mut TabList._buffer,
-      c_int::from(TabList.NameOffsets[id] - 1),
+      c_int::from(TabList.NameOffsets[id as usize] - 1),
     )
   }
   .to_string()
